Test on both Linux and OSX.
Switch to C language so we can specify operating systems. We have to install ruby 2.1 via rvm because the ruby interpreter on travis ci crashes when running rake ci, however that same version of ruby works fine on my local machine. This work around adds about 6 Minutes to the build and should be replaced as soon as possible to speed up builds.